templates/python-rest-api/requirements.txt
fastapi==0.116.1
python-dotenv==1.1.1
requests==2.32.5
python-multipart==0.0.20
uvicorn==0.35.0

